worker/requirements.txt
# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17

# Database
motor==3.6.0

# AI/ML
google-generativeai==0.8.3
langchain==0.3.10
langchain-google-genai==2.0.7

# Audio Processing
pydub==0.25.1
gTTS==2.5.4
librosa==0.10.2.post1
soundfile==0.12.1

# Image Processing
Pillow==11.0.0
opencv-python==4.10.0.84
numpy==1.26.4

# AWS
boto3==1.35.72

# Utils
python-dotenv==1.0.1
pydantic==2.10.3
httpx==0.28.1
aiofiles==24.1.0

# Development
pytest==8.3.4
pytest-asyncio==0.25.0